refactor: extract pure YouTube ingest formatting into server/lib (#6015)
`server/services/youtubeIngest.js` mixed six responsibilities in one 865-line
file, so unit-testing YAML escaping or the CoS agent prompt meant importing the
whole service graph (child process, brain/cos/media stores, fs) behind a wall of
`vi.mock` stubs.

Move the pure half — `parseVideoMetadata`, `buildIngestNote`,
`buildAgentTaskContext`, `resolveObsidianPointer`, `sanitizeFilename`,
`formatDuration`, `yamlString` — verbatim into the new
`server/lib/youtubeIngestFormat.js`. The moved bodies are byte-identical; the
service imports them and keeps orchestration, storage, and the yt-dlp spawn.

The new module is surfaced through the `server/lib` barrel as a NAMESPACE export
(`youtubeIngestFormat.*`) rather than a flat one: `formatDuration` collides with
`fileCore.js` and `sanitizeFilename` with `mimeTypes.js`, which the barrel's
collision guard rejects for flat `export *`.

Tests for those four public contracts move to
`server/lib/youtubeIngestFormat.test.js` and now run with no mocks at all — if
that file ever needs one, something impure leaked back into lib/. The service
suite keeps the URL-allowlist and cancel-escalation cases. Added coverage for
the untrusted-transcript boundary notice in the agent prompt, which nothing
pinned before.
